added additional route for authorization
and changed some response on authorization
