+++ app.js
@@ -21,6 +21,7 @@ db.once('open', function(){
 
 var routes = require('./routes/index');
 var users = require('./routes/users');
+var authorization = require('./routes/authorization');
 
 var app = express();
 
@@ -44,6 +45,7 @@ app.use(express.static(path.join(__dirname, 'public')));
 
 app.use('/', routes);
 app.use('/users', users);
+app.use('/auth', authorization);
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
